openssl: fix compiler warning with BoringSSL/AWS-LC with debug logging enabled
This API returns `uint32_t` with these two forks, and `unsigned long`
elsewhere.

```
src/openssl.c: In function ‘ssh2_ed25519_verify’:
src/openssl.c:3291:19: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 5 has type ‘uint32_t’ {aka ‘unsigned int’} [-Werror=format=]
 3291 |                   "ssh2_ed25519_verify(): EVP_DigestVerify()->%d (ossl: %lu)",
      |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 3292 |                   ret, ret ? ERR_peek_last_error() : 0));
      |                        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |                                                    |
      |                                                    uint32_t {aka unsigned int}
src/libssh2_priv.h:1112:34: note: in definition of macro ‘ssh2_deb’
```
Ref: https://github.com/libssh2/libssh2/actions/runs/30837337514/job/91765629453?pr=2503#step:24:105

+++ src/openssl.c
@@ -3289,7 +3289,7 @@ int ssh2_ed25519_verify(ssh2_ed25519_ctx *ed_ctx, LIBSSH2_SESSION *session,
     if(ret != 1)
         ssh2_deb((session, LIBSSH2_TRACE_KEX,
                   "ssh2_ed25519_verify(): EVP_DigestVerify()->%d (ossl: %lu)",
-                  ret, ret ? ERR_peek_last_error() : 0));
+                  ret, ret ? (unsigned long)ERR_peek_last_error() : 0));
 
 clean_exit:
